refactor rpc files to have _rpc suffix to remove ambigious clippy error
