+++ newsfragments/690.fixed.md
@@ -1,0 +1,1 @@
+Refactor rpc files to have _rpc suffix to remove ambigious clippy error when using discv5
