+++ rpc/src/history_rpc.rs
