+++ rpc/src/lib.rs
@@ -1,12 +1,12 @@
 #![warn(clippy::unwrap_used)]
 
-mod discv5;
-mod history;
-mod server;
-mod web3;
+mod discv5_rpc;
+mod history_rpc;
+mod server_rpc;
+mod web3_rpc;
 
-pub use discv5::Discv5Api;
+pub use discv5_rpc::Discv5Api;
 pub use ethportal_api::jsonrpsee;
-pub use history::HistoryNetworkApi;
-pub use server::JsonRpcServer;
-pub use web3::Web3Api;
+pub use history_rpc::HistoryNetworkApi;
+pub use server_rpc::JsonRpcServer;
+pub use web3_rpc::Web3Api;
